refactor(cliproxy): extract smoke test runner into setup/smoke-test.ts
Move SmokeResult, GhRunEntry, SmokeTestInternals, and runSmokeTest into a
focused submodule. Drops the dead second branch of the environment-approval
gate (status='pending' + /approval/.test(conclusion) was unreachable because
gh returns null conclusion while pending). Adds an R5/4f documentation test
locking in the known race-attribution limitation when concurrent contributors
trigger workflow runs.

setup.ts shrinks from 758 to 555 LOC. Adds 2 new TDD-first tests (722 total, +2).
